refactor config and properties. Cleanup some injection points.
