+++ state-conductor-dataservices/build.gradle
@@ -32,8 +32,8 @@ dependencies {
   implementation 'com.marklogic:ml-javaclient-util:3.13.4'
   implementation 'com.google.guava:guava:28.1-jre'
 
-  implementation "org.springframework.boot:spring-boot-starter:${springVersion}"
-  //implementation "org.springframework.boot:spring-boot-starter-web:${springVersion}"
+  //implementation "org.springframework.boot:spring-boot-starter:${springVersion}"
+  implementation "org.springframework.boot:spring-boot-starter-web:${springVersion}"
   testImplementation("org.springframework.boot:spring-boot-starter-test:${springVersion}") {
     exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
   }
@@ -50,20 +50,20 @@ dependencies {
 	runtimeOnly group: "org.slf4j", name: "slf4j-api", version: "1.7.22"
 }
 
-mainClassName = 'com.marklogic.StateConductorDriver'
+//mainClassName = 'com.marklogic.stateconductor.StateConductorDriver'
 sourceCompatibility = 1.8
 targetCompatibility = 1.8
 
-jar {
-  manifest {
-    attributes(
-      'Main-Class': 'com.marklogic.stateconductor.StateConductorDriver'
-    )
-  }
-  from {
-    configurations.runtimeClasspath.collect { it.isDirectory() ? it : zipTree(it) }
-  }
-}
+//jar {
+//  manifest {
+//    attributes(
+//      'Main-Class': 'com.marklogic.stateconductor.StateConductorDriver'
+//    )
+//  }
+//  from {
+//    configurations.runtimeClasspath.collect { it.isDirectory() ? it : zipTree(it) }
+//  }
+//}
 
 task generateStateConductorProxy(type: com.marklogic.client.tools.gradle.EndpointProxiesGenTask) {
   dependsOn 'mlPrepareBundles'
