+++ state-conductor-dataservices/public/index.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <title>Hello World</title>
+</head>
+<body>
+<h1>Hello Spring!</h1>
+</body>
+</html>
