+++ state-conductor-dataservices/src/main/resources/application.properties
@@ -22,3 +22,4 @@ fixedThreadCount=-1
 threadsPerHost=16
 maxThreadCount=128
 
+server.port=9000
